+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Change Log
 
 ## Unreleased
+
+### Fixed
+
+- avoid error on watch non `tf.Tensor` in tensorflow backend grad method
+
+- circuit preprocessing simplification with only single qubit gates